build: switch to npm (#772)
#### Why
Moving to NPM to be more consistent across LINZ, most of LINZ uses npm
and splitting the codebases between NPM and yarn is confusing and time
consuming. Also the version of yarn we are using is deprecated and hasnt
see significant improvements recently, where as npm has started getting
a lot of new features.