config: add a label base style BM-1066 (#966)
### Motivation

It would be nice to layer all the vector data ontop of all the
individual aerial imagery datasets that we have.

### Modifications

Add a "labels" style which has no base layer and overlays labels and
roads.

### Verification

When it is deployed we can verify it looks good across all the
individual layers.