chore: Update CODEOWNERS (#1068)
### Motivation

Asked the Elevation team to review a change to their configs and they
couldn't!

### Modifications

Adjust the CODEOWNERS file to account for recent config additions and
alterations, to ensure the imagery and elevation data maintainers have
review rights on the config files associated to their data.

### Verification

Check if they changes work after...