ci: Upload changes.md as artifact for oversized comments. (#1334)
### Motivation

Sometime when we adding lots of new layers in one PR, the changed.md
might exceed the limitation of the single PR comment size. See
[example](https://github.com/linz/basemaps-config/actions/runs/17083214295/job/48441904227)

This ci changes will upload the changed.md file as artifact and provide
a download link the comment that point to it.
As well as provide a simple comment with the preview links of the PR
config parameter.

### Modifications

- Add a step to upload changed.md if exists
- Check file size and update the comment if oversized.

### Verification

Works well in the charts
[PR](https://github.com/linz/basemaps-config/pull/1231#issuecomment-3384031170).