feat: increase resolution of some sprites BM-350 (#1342)
### Motivation

When users are browsing topographic styles in a standard DPI device they
recieve the `@1x` spritesheet. This topographic style has a number of
rules where the sprites are rendered larger than the are stored in the
1x sprite sheet.

For instance, cliff_edge is stored as a 32x16 sprite then rendered as a
20px line, when maplibre see this it upscales the sprite causing it to
become blury

When users use the highdpi sprites the sprite is stored as 64x32 so it
is scaled down to 20px which allows it to remain crisp.


### Modifications

Increase the resolution of sprites that are scaled up by the line-width
property in the topographic style.

<!-- TODO: Say what changes you made. -->

<!-- TODO: Attach screenshots if you changed the UI. -->

### Verification

<!-- TODO: Say how you tested your changes. -->
Screenshot diffing shows a remarkable improvement in the sprite
rendering and lowdpi


https://d25mfjh9syaxsr.cloudfront.net/linz/basemaps-config/2025/10/27-adc796ff29e884798441e076e232e7f8d466d633/index.html